refactor(fullscreen-wrapper): move screen to drawer navigator
